Move station heartbeat to its own file
